+++ sidekick/lib/sidekick.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
-import 'package:sidekick/src/init/init_command.dart';
-import 'package:sidekick/src/plugins/plugins_command.dart';
+import 'package:sidekick/src/commands/init_command.dart';
+import 'package:sidekick/src/commands/plugins_command.dart';
 import 'package:sidekick_core/sidekick_core.dart' as core;
 import 'package:sidekick_core/sidekick_core.dart' show Version;
