+++ sidekick/lib/src/commands/init_command.dart
@@ -1,8 +1,8 @@
 import 'package:dcli/dcli.dart' as dcli;
 import 'package:recase/recase.dart';
-import 'package:sidekick/src/init/name_suggester.dart';
 import 'package:sidekick/src/util/dcli_ask_validators.dart';
 import 'package:sidekick/src/util/directory_extension.dart';
+import 'package:sidekick/src/util/name_suggester.dart';
 import 'package:sidekick_core/sidekick_core.dart';
 
 class InitCommand extends Command {
