+++ sidekick/lib/src/commands/plugins_command.dart
