+++ sidekick/lib/src/util/name_suggester.dart
